@@ -6,11 +6,11 @@
     "authors": [
         {
             "name": "Leon Dietsch",
-            "email": "leon.dietsch@netzleuchten.com"
+            "email": "hallo@leondietsch.de"
         },
         {
             "name": "Holger McCloy",
-            "email": "holger.mccloy@netzleuchten.com"
+            "email": "hallo@netzleuchten.com"
         }
     ],
     "scripts": {
@@ -31,7 +31,7 @@
         }
     },
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.1",
         "php-di/php-di": "^5.4",
         "doctrine/cache": "^1.6",
         "guzzlehttp/guzzle": "^6.2",
@@ -42,8 +42,8 @@
     "require-dev": {
         "phpunit/phpunit": "^6.0",
         "hmaus/drafter-php": "^4.0",
-        "phpmetrics/phpmetrics": "^2.0",
-        "squizlabs/php_codesniffer": "3.0.0RC4",
-        "phpmd/phpmd": "^2.6"
+        "phpmetrics/phpmetrics": "^2.5",
+        "squizlabs/php_codesniffer": "^3.5",
+        "phpmd/phpmd": "^2.8"
     }
 }
