@@ -2,7 +2,12 @@
 
 namespace Nordkirche\Ndk\Service;
 
-class ResolutionProxy
+use Nordkirche\Ndk\Domain\Interfaces\ModelInterface;
+use Nordkirche\Ndk\Domain\Model\ResourcePlaceholder;
+use Nordkirche\Ndk\Service\Exception\NdkException;
+use Nordkirche\Ndk\Service\Interfaces\ProxyInterface;
+
+class ResolutionProxy implements ProxyInterface
 {
     /**
      * @var string
@@ -31,10 +36,30 @@ class ResolutionProxy
     }
 
     /**
-     * @return \Nordkirche\Ndk\Domain\Interfaces\ModelInterface|Result|null
+     * @return ModelInterface|Result|null
      */
     public function resolve()
     {
-        return $this->napiService->get($this->uri);
+        try {
+            return $this->napiService->get($this->uri);
+        } catch (Exception\ResourceObjectNotFoundException $e) {
+            return $this->buildResourcePlaceholder($e);
+        } catch (Exception\TimeoutException $e) {
+            return $this->buildResourcePlaceholder($e);
+        } catch (Exception\ServerException $e) {
+            return $this->buildResourcePlaceholder($e);
+        }
+    }
+
+    private function buildResourcePlaceholder(NdkException $exception): ResourcePlaceholder
+    {
+        [$type, $id] = $this->napiService->returnTypeAndIdFromPath($this->uri);
+
+        $placeholder = new ResourcePlaceholder();
+        $placeholder->setLinkedException($exception);
+        $placeholder->setId($id);
+        $placeholder->setType($type);
+
+        return $placeholder;
     }
 }
