@@ -2,6 +2,7 @@
 
 namespace Nordkirche\Ndk\Service;
 
+use Monolog\Logger;
 use Nordkirche\Ndk\Helper\AbstractIntegrationTestCase;
 
 class ResolutionServiceTest extends AbstractIntegrationTestCase
@@ -10,7 +11,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
     /**
      * @var ResolutionService
      */
-    private $testSubject;
+    private $resolutionService;
 
     private $singelPrimaryData = [
         'id' => 1,
@@ -94,6 +95,32 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         ]
     ];
 
+    private $skipRelationshipWhenIncludeIsMissing = [
+        'data' => [
+            'id' => '1',
+            'type' => 'mock-model-type',
+            'attributes' => [
+                'integer' => 1234,
+                'string' => 'Some nice string for you.'
+            ],
+            'relationships' => [
+                'object' => [
+                    'links' => [],
+                    'data' => [
+                        'id' => '2',
+                        'type' => 'mock-model-sibling-type',
+                    ]
+                ]
+            ]
+        ],
+        'included' => [
+        ],
+        'meta' => [
+            'page_count' => 1,
+            'record_count' => 1
+        ]
+    ];
+
     private $resolveIncludeDataMultipleRelations = [
         'data' => [
             'id' => '1',
@@ -129,6 +156,58 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
             ],
             [
                 'id' => '3',
+                'type' => 'mock-model-sibling-type',
+                'attributes' => [
+                    'integer' => '1234',
+                    'string' => 'Some nice string for you.'
+                ]
+            ]
+        ],
+        'meta' => [
+            'page_count' => 1,
+            'record_count' => 1
+        ]
+    ];
+
+    private $skipSubRelationshipsWhenIncludeIsMissing = [
+        'data' => [
+            'id' => '1',
+            'type' => 'mock-model-type',
+            'attributes' => [
+                'integer' => 1234,
+                'string' => 'Some nice string for you.'
+            ],
+            'relationships' => [
+                'resultObject' => [
+                    'links' => [],
+                    'data' => [
+                        [
+                            'id' => '2',
+                            'type' => 'mock-model-sibling-type',
+                        ],
+                        [
+                            'id' => '3',
+                            'type' => 'mock-model-sibling-type',
+                        ],
+                        [
+                            'id' => '4',
+                            'type' => 'mock-model-sibling-type',
+                        ]
+                    ]
+                ]
+            ]
+        ],
+        'included' => [
+            [
+                'id' => '2',
+                'type' => 'mock-model-sibling-type',
+                'attributes' => [
+                    'integer' => '1234',
+                    'string' => 'Some nice string for you.'
+                ]
+            ],
+            [
+                'id' => '4',
                 'type' => 'mock-model-sibling-type',
                 'attributes' => [
                     'integer' => '1234',
@@ -200,7 +279,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
             'mock-model-type' => \Nordkirche\Ndk\Service\MockModel::class,
             'mock-model-sibling-type' => \Nordkirche\Ndk\Service\MockSiblingModel::class
         ]);
-        $this->testSubject = $this->api->factory(ResolutionService::class);
+        $this->resolutionService = $this->api->factory(ResolutionService::class);
     }
 
     /**
@@ -211,7 +290,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
         $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
 
-        $this->testSubject->resolve(
+        $this->resolutionService->resolve(
             new \Nordkirche\Ndk\Service\Response(
                 $mockResponse,
                 [
@@ -227,9 +306,9 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var MockModel $object */
         for ($i = 0; $i < 2; $i++) {
 
-            $object = $this->testSubject->get('mock-model-type', $this->includedData[$i]['id']);
+            $object = $this->resolutionService->get('mock-model-type', $this->includedData[$i]['id']);
 
-            self::assertTrue($object instanceof MockModel, 'Wrong object was created.');
+            self::assertTrue($object instanceof MockModel, 'Wrong object was created');
             self::assertEquals($object->getId(), $this->includedData[$i]['id'], 'Wrong value');
             self::assertEquals($object->getType(), $this->includedData[$i]['type'], 'Wrong value');
             self::assertEquals($object->getInteger(), $this->includedData[$i]['attributes']['integer'], 'Wrong value');
@@ -245,7 +324,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
         $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
 
-        $this->testSubject->resolve(
+        $this->resolutionService->resolve(
             new \Nordkirche\Ndk\Service\Response(
                 $mockResponse,
                 $this->searchData
@@ -253,9 +332,9 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         );
 
         /** @var MockModel $object */
-        $object = $this->testSubject->get('mock-model-type', 1);
+        $object = $this->resolutionService->get('mock-model-type', 1);
 
-        self::assertTrue($object instanceof MockModel, 'Wrong object was created.');
+        self::assertTrue($object instanceof MockModel, 'Wrong object was created');
 
         self::assertInstanceOf(\Nordkirche\Ndk\Domain\Model\MetaData::class, $object->getMetaData());
         self::assertEquals($this->searchData['data'][0]['meta']['score'], $object->getMetaData()->getScore());
@@ -269,7 +348,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
         $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
 
-        $this->testSubject->resolve(
+        $this->resolutionService->resolve(
             new \Nordkirche\Ndk\Service\Response(
                 $mockResponse,
                 $this->resolveIncludeDataMultipleRelations
@@ -277,11 +356,12 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         );
 
         /** @var MockModel $object */
-        $object = $this->testSubject->get('mock-model-type', 1);
+        $object = $this->resolutionService->get('mock-model-type', 1);
 
-        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created.');
+        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created');
         self::assertInstanceOf(Result::class, $object->getResultObject(), 'Result object is missing.');
-        self::assertInstanceOf(MockSiblingModel::class, $object->getResultObject()->current(), 'Child object is missing.');
+        self::assertInstanceOf(MockSiblingModel::class, $object->getResultObject()->current(),
+            'Child object is missing.');
     }
 
     /**
@@ -292,7 +372,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
         $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
 
-        $this->testSubject->resolve(
+        $this->resolutionService->resolve(
             new \Nordkirche\Ndk\Service\Response(
                 $mockResponse,
                 $this->resolveIncludeData
@@ -300,10 +380,101 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         );
 
         /** @var MockModel $object */
-        $object = $this->testSubject->get('mock-model-type', 1);
+        $object = $this->resolutionService->get('mock-model-type', 1);
 
-        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created.');
-        self::assertInstanceOf(MockSiblingModel::class, $object->getObject(), 'Wrong subobject 1st level was created.');
+        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created');
+        self::assertInstanceOf(MockSiblingModel::class, $object->getObjectRaw(), 'Wrong subobject 1st level was created.');
+    }
+
+    /**
+     * @group integration
+     */
+    public function testReturnResolutionProxyWhenIncludeIsMissing()
+    {
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipRelationshipWhenIncludeIsMissing
+            )
+        );
+
+        /** @var MockModel $object */
+        $object = $this->resolutionService->get('mock-model-type', 1);
+
+        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created');
+        self::assertInstanceOf(ResolutionProxy::class, $object->getObjectRaw(), 'Wrong subobject 1st level was created.');
+    }
+
+    /**
+     * @group integration
+     */
+    public function testSkipRelationShipWhenIncludeIsMissing()
+    {
+        $this->api->getConfiguration()->setResolutionProxyDisabled(true);
+
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipRelationshipWhenIncludeIsMissing
+            )
+        );
+
+        /** @var MockModel $object */
+        $object = $this->resolutionService->get('mock-model-type', 1);
+
+        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created');
+        self::assertInstanceOf(MissingResourceProxy::class, $object->getObjectRaw(), 'Wrong object was created');
+        self::assertNull($object->getObject());
+    }
+
+    /**
+     * @group integration
+     */
+    public function testLogSkipRelationshipWhenIncludeIsMissing()
+    {
+        $this->api->getConfiguration()->setResolutionProxyDisabled(true);
+
+        $handler = $this->setUpLogger();
+
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipRelationshipWhenIncludeIsMissing
+            )
+        );
+
+        /** @var MockModel $mockModel */
+        $mockModel = $this->resolutionService->get('mock-model-type', 1);
+
+        $object = $mockModel->getObjectRaw();
+
+        self::assertInstanceOf(MissingResourceProxy::class, $object);
+
+        $mockModel->getObject(); // this should trigger a log entry
+
+        $log = $handler->close();
+
+        self::assertNotEmpty($log);
+        self::assertCount(1, $log);
+
+        $logEntry = $log[0];
+        self::assertArrayHasKey('message', $logEntry);
+        self::assertEquals($logEntry['message'], 'Skipped relationship because it was not found in includes');
+
+        self::assertArrayHasKey('context', $logEntry);
+        self::assertEquals($logEntry['context'], ['type' => 'mock-model-sibling-type', 'id' => '2']);
+
+        self::assertArrayHasKey('level', $logEntry);
+        self::assertEquals($logEntry['level'], Logger::INFO);
     }
 
     /**
@@ -314,7 +485,7 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
         $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
 
-        $this->testSubject->resolve(
+        $this->resolutionService->resolve(
             new \Nordkirche\Ndk\Service\Response(
                 $mockResponse,
                 $this->resolveSubIncludeData
@@ -322,11 +493,11 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
         );
 
         /** @var MockModel $object */
-        $object = $this->testSubject->get('mock-model-type', 1);
+        $object = $this->resolutionService->get('mock-model-type', 1);
 
-        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created.');
-        self::assertInstanceOf(MockSiblingModel::class, $object->getObject(), 'Wrong subobject 1st level was created.');
-        self::assertInstanceOf(MockSiblingModel::class, $object->getObject()->getObject(),
+        self::assertInstanceOf(MockModel::class, $object, 'Wrong object was created');
+        self::assertInstanceOf(MockSiblingModel::class, $object->getObjectRaw(), 'Wrong subobject 1st level was created.');
+        self::assertInstanceOf(MockSiblingModel::class, $object->getObjectRaw()->getObject(),
             'Wrong subobject 2nd level was created.');
     }
 
@@ -345,14 +516,123 @@ class ResolutionServiceTest extends AbstractIntegrationTestCase
             ]
         );
 
-        $this->testSubject->resolve($response);
+        $this->resolutionService->resolve($response);
 
-        self::assertTrue($response->getPrimaryData() instanceof MockModel, 'Wrong object was created.');
+        self::assertTrue($response->getPrimaryData() instanceof MockModel, 'Wrong object was created');
         self::assertEquals($response->getPrimaryData()->getId(), $this->singelPrimaryData['id'], 'Wrong value');
         self::assertEquals($response->getPrimaryData()->getType(), $this->singelPrimaryData['type'], 'Wrong value');
         self::assertEquals($response->getPrimaryData()->getInteger(), $this->singelPrimaryData['attributes']['integer'],
             'Wrong value');
         self::assertEquals($response->getPrimaryData()->getString(), $this->singelPrimaryData['attributes']['string'],
             'Wrong value');
+    }
+
+    /**
+     * @group integration
+     */
+    public function testReturnResolutionProxyWhenSubIncludeIsMissing()
+    {
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipSubRelationshipsWhenIncludeIsMissing
+            )
+        );
+
+        /** @var MockModel $object */
+        for ($i = 0; $i < 2; $i++) {
+            $object = $this->resolutionService->get('mock-model-type', '1');
+
+            self::assertTrue($object instanceof MockModel, 'Wrong object was created: ' . get_class($object));
+
+            $resultObject = $object->getResultObject();
+
+            self::assertEquals(3, $resultObject->count(), 'Resolved too many objects');
+
+            /** @var MockSiblingModel $mockSibling */
+            foreach ($resultObject as $i => $mockSibling) {
+                if ($i === 1) {
+                    self::assertInstanceOf(ResolutionProxy::class, $mockSibling);
+                } else {
+                    self::assertInstanceOf(MockSiblingModel::class, $mockSibling);
+                }
+            }
+        }
+    }
+
+    /**
+     * @group integration
+     */
+    public function testSkipMissingSubIncludes()
+    {
+        $this->api->getConfiguration()->setResolutionProxyDisabled(true);
+
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipSubRelationshipsWhenIncludeIsMissing
+            )
+        );
+
+        /** @var MockModel $object */
+        for ($i = 0; $i < 2; $i++) {
+            $object = $this->resolutionService->get('mock-model-type', '1');
+
+            self::assertTrue($object instanceof MockModel, 'Wrong object was created: ' . get_class($object));
+
+            $resultObject = $object->getResultObject();
+
+            self::assertEquals(2, $resultObject->count(), 'Resolved too many objects');
+
+            /** @var MockSiblingModel $mockSibling */
+            foreach ($resultObject as $mockSibling) {
+                if ($mockSibling->getId() === '3') {
+                    self::fail('Relationship with id 3 should never have been resolved');
+                } else {
+                    self::assertInstanceOf(MockSiblingModel::class, $mockSibling);
+                }
+            }
+        }
+    }
+
+    /**
+     * @group integration
+     */
+    public function testLogSkippedMissingSubIncludes()
+    {
+        $this->api->getConfiguration()->setResolutionProxyDisabled(true);
+
+        $handler = $this->setUpLogger();
+
+        /** @var \Psr\Http\Message\ResponseInterface $mockResponse */
+        $mockResponse = $this->createMock(\Psr\Http\Message\ResponseInterface::class);
+
+        $this->resolutionService->resolve(
+            new \Nordkirche\Ndk\Service\Response(
+                $mockResponse,
+                $this->skipSubRelationshipsWhenIncludeIsMissing
+            )
+        );
+
+        $log = $handler->close();
+
+        self::assertNotEmpty($log);
+        self::assertCount(1, $log);
+
+        $logEntry = $log[0];
+        self::assertArrayHasKey('message', $logEntry);
+        self::assertEquals($logEntry['message'], 'Skipped relationship because it was not found in includes');
+
+        self::assertArrayHasKey('context', $logEntry);
+        self::assertEquals($logEntry['context'], ['type' => 'mock-model-sibling-type', 'id' => '3']);
+
+        self::assertArrayHasKey('level', $logEntry);
+        self::assertEquals($logEntry['level'], Logger::INFO);
     }
 }
